Rename Intake.java to RollerIntake.java
