+++ src/main/java/frc/robot/commands/scoring/intake/IntakeGamePiece.java
@@ -2,20 +2,20 @@ package frc.robot.commands.scoring.intake;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.constants.IntakeConstants;
-import frc.robot.subsystems.Intake;
-import frc.robot.subsystems.Intake.IntakeMode;
-import frc.robot.subsystems.Intake.IntakePiece;
+import frc.robot.subsystems.RollerIntake;
+import frc.robot.subsystems.RollerIntake.IntakeMode;
+import frc.robot.subsystems.RollerIntake.IntakePiece;
 
 public class IntakeGamePiece extends CommandBase {
 
-  private final Intake m_intake; 
+  private final RollerIntake m_intake; 
   private final IntakePiece m_type;
 
   /**
    * Spins the intake until the game piece is inside the intake.
    * @param intake the intake subsystem
    */
-  public IntakeGamePiece(Intake intake, IntakePiece type) {
+  public IntakeGamePiece(RollerIntake intake, IntakePiece type) {
     m_intake = intake; 
     m_type = type;
     addRequirements(m_intake);
