+++ src/main/java/frc/robot/controls/ManualController.java
@@ -7,7 +7,7 @@ import frc.robot.constants.OIConstants;
 import frc.robot.subsystems.Elevator;
 import frc.robot.subsystems.Elevator.ElevatorMode;
 import frc.robot.subsystems.FourBarArm;
-import frc.robot.subsystems.Intake;
+import frc.robot.subsystems.RollerIntake;
 import frc.robot.util.Functions;
 import lib.controllers.GameController;
 import lib.controllers.GameController.Axis;
@@ -17,10 +17,10 @@ import lib.controllers.GameController.DPad;
 public class ManualController {
   GameController m_manual = new GameController(OIConstants.kManualJoy);
   private FourBarArm m_arm;
-  private Intake m_intake;
+  private RollerIntake m_intake;
   private Elevator m_elevator;
   
-  public ManualController(FourBarArm arm, Intake intake, Elevator elevator) {
+  public ManualController(FourBarArm arm, RollerIntake intake, Elevator elevator) {
     m_arm = arm;
     m_intake = intake;
     m_elevator = elevator;
